refactor(aztec-node): extract tx receipt assembly into a module
Move getTxReceipt and its private helpers (assembleMinedReceipt, deriveMinedStatus) into a new
`NodeTxReceiptBuilder` collaborator under `src/modules/`, built over the tx pool, block source, and debug
log store. `AztecNodeService` keeps a thin delegating getTxReceipt. server.ts is now ~860 lines, down from
~2140 at the start of this refactor.
