+++ yarn-project/aztec-node/src/aztec-node/server.ts
@@ -49,7 +49,6 @@ import type {
   NodeInfo,
   ProtocolContractAddresses,
 } from '@aztec/stdlib/contract';
-import { getEpochAtSlot } from '@aztec/stdlib/epoch-helpers';
 import { GasFees, type ManaUsageEstimate, getNetworkTxGasLimits } from '@aztec/stdlib/gas';
 import type {
   AztecNode,
@@ -84,20 +83,15 @@ import type { CheckpointAttestation } from '@aztec/stdlib/p2p';
 import type { Offense } from '@aztec/stdlib/slashing';
 import { MerkleTreeId, NullifierMembershipWitness, PublicDataWitness } from '@aztec/stdlib/trees';
 import {
-  DroppedTxReceipt,
   type FeeProvider,
   type GetTxReceiptOptions,
   type GlobalVariableBuilder as GlobalVariableBuilderInterface,
   type IndexedTxEffect,
-  MinedTxReceipt,
-  type MinedTxStatus,
-  PendingTxReceipt,
   PublicSimulationOutput,
   type SimulationOverrides,
   Tx,
   type TxHash,
   type TxReceipt,
-  TxStatus,
   type TxValidationResult,
 } from '@aztec/stdlib/tx';
 import type { SingleValidatorStats, ValidatorsStats } from '@aztec/stdlib/validators';
@@ -112,6 +106,7 @@ import {
 import { NodeKeystoreAdapter, ValidatorClient } from '@aztec/validator-client';
 
 import { NodeBlockProvider } from '../modules/node_block_provider.js';
+import { NodeTxReceiptBuilder } from '../modules/node_tx_receipt.js';
 import { NodeWorldStateQueries } from '../modules/node_world_state_queries.js';
 import { Sentinel } from '../sentinel/sentinel.js';
 import type { AztecNodeConfig } from './config.js';
@@ -166,6 +161,7 @@ export class AztecNodeService implements AztecNode, AztecNodeAdmin, AztecNodeDeb
   private readonly nodePublicCallsSimulator: NodePublicCallsSimulator;
   private readonly worldStateQueries: NodeWorldStateQueries;
   private readonly blockProvider: NodeBlockProvider;
+  private readonly txReceiptBuilder: NodeTxReceiptBuilder;
 
   public readonly tracer: Tracer;
 
@@ -255,6 +251,12 @@ export class AztecNodeService implements AztecNode, AztecNodeAdmin, AztecNodeDeb
     });
 
     this.blockProvider = new NodeBlockProvider(this.blockSource);
+
+    this.txReceiptBuilder = new NodeTxReceiptBuilder({
+      p2pClient: this.p2pClient,
+      blockSource: this.blockSource,
+      debugLogStore: this.debugLogStore,
+    });
 
     this.log.info(`Aztec Node version: ${this.packageVersion}`);
     this.log.info(`Aztec Node started on chain 0x${this.l1ChainId.toString(16)}`, pickL1ContractAddresses(this.config));
@@ -530,78 +532,11 @@ export class AztecNodeService implements AztecNode, AztecNodeAdmin, AztecNodeDeb
     this.log.info(`Received tx ${txHash} in ${duration}ms`, { txHash });
   }
 
-  public async getTxReceipt<TGetTxReceiptOptions extends GetTxReceiptOptions = {}>(
+  public getTxReceipt<TGetTxReceiptOptions extends GetTxReceiptOptions = {}>(
     txHash: TxHash,
     options?: TGetTxReceiptOptions,
   ): Promise<TxReceipt<TGetTxReceiptOptions>> {
-    // Check the tx pool status first. If the tx is known to the pool (pending or mined), we'll use that
-    // as a fallback if we don't find a mined tx effect in the archiver.
-    const txPoolStatus = await this.p2pClient.getTxStatus(txHash);
-    const isKnownToPool = txPoolStatus === 'pending' || txPoolStatus === 'mined';
-
-    // Then get the raw tx effect from the archiver, which tracks every tx in a mined block.
-    const indexed = await this.blockSource.getTxEffect(txHash);
-
-    let receipt: TxReceipt;
-    if (indexed) {
-      receipt = await this.#assembleMinedReceipt(indexed, options);
-    } else if (isKnownToPool) {
-      // If the tx is in the pool but not in the archiver, it's pending.
-      // This handles race conditions between archiver and p2p, where the archiver
-      // has pruned the block in which a tx was mined, but p2p has not caught up yet.
-      let tx: Tx | undefined;
-      if (options?.includePendingTx) {
-        // The tx may have left the pool since we checked its status (mined or dropped); in that case we
-        // leave `tx` unset and still return a pending receipt.
-        tx = await this.p2pClient.getTxByHashFromPool(txHash, { includeProof: !!options.includeProof });
-      }
-      receipt = new PendingTxReceipt(txHash, tx);
-    } else {
-      // Otherwise, if we don't know the tx, we consider it dropped.
-      receipt = new DroppedTxReceipt(txHash, 'Tx dropped by P2P node');
-    }
-
-    this.debugLogStore.decorateReceiptWithLogs(txHash.toString(), receipt);
-
-    return receipt;
-  }
-
-  /**
-   * Assembles a {@link MinedTxReceipt} from a raw {@link IndexedTxEffect}, deriving the finalization status from the
-   * cached L2 tips and the epoch from the block's slot number.
-   */
-  async #assembleMinedReceipt(indexed: IndexedTxEffect, options?: GetTxReceiptOptions): Promise<MinedTxReceipt> {
-    const blockNumber = indexed.l2BlockNumber;
-    const [tips, l1Constants] = await Promise.all([this.blockSource.getL2Tips(), this.blockSource.getL1Constants()]);
-
-    const status = this.#deriveMinedStatus(blockNumber, tips);
-    const epochNumber = getEpochAtSlot(indexed.slotNumber, l1Constants);
-
-    return new MinedTxReceipt(
-      indexed.data.txHash,
-      status,
-      MinedTxReceipt.executionResultFromRevertCode(indexed.data.revertCode),
-      indexed.data.transactionFee.toBigInt(),
-      indexed.l2BlockHash,
-      blockNumber,
-      indexed.slotNumber,
-      indexed.txIndexInBlock,
-      epochNumber,
-      options?.includeTxEffect ? indexed.data : undefined,
-      /*debugLogs=*/ undefined,
-    );
-  }
-
-  #deriveMinedStatus(blockNumber: BlockNumber, tips: L2Tips): MinedTxStatus {
-    if (blockNumber <= tips.finalized.block.number) {
-      return TxStatus.FINALIZED;
-    } else if (blockNumber <= tips.proven.block.number) {
-      return TxStatus.PROVEN;
-    } else if (blockNumber <= tips.checkpointed.block.number) {
-      return TxStatus.CHECKPOINTED;
-    } else {
-      return TxStatus.PROPOSED;
-    }
+    return this.txReceiptBuilder.getTxReceipt(txHash, options);
   }
 
   public getTxEffect(txHash: TxHash): Promise<IndexedTxEffect | undefined> {
